edit

+++ function_pointers/1-array_iterator.c
@@ -1,5 +1,6 @@
 #include "function_pointers.h"
 #include <stdlib.h>
+#include <stddef.h>
 
 /**
 * array_iterator - Executes a function on each elements of an array
@@ -12,7 +13,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i = 0;
+	size_t i = 0;
 
 	if (array == NULL || action == NULL)
 		return;
